オプションなしの場合の整形作業をクラスに分ける
