+++ 07.ls_object/formatter.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class Formatter
+  def initialize(files)
+    @files = files
+  end
+
+  def self.find_longest_string_length(file_prop_strings)
+    file_prop_strings.map(&:size).max
+  end
+
+  def format_files; end
+end
